Lista de Usuarios

+++ app/src/main/java/com/example/firebasenotes/WidgetsCardView/Listing/ListingDrawer/Registre.kt
@@ -113,7 +113,7 @@ fun ComponentRegistrar(drawerViewModel: DrawerViewModel = viewModel()) {
 
 
         Button(
-            onClick = { drawerViewModel.insertarDatos(num_cajon,nombre_cajon ,piso_edificio,selectedOptionText
+            onClick = { drawerViewModel.insertarDatos(num_cajon.toInt(),nombre_cajon ,piso_edificio,selectedOptionText
             ,desc,img,esEspecial = false)
                       num_cajon = ""
                       nombre_cajon = ""
